+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.1.0] - 2026-07-27
+
 ### Added
 - Air Quality threshold config fields (`airQualityExcellentMaxPpm`,
   `airQualityGoodMaxPpm`, `airQualityFairMaxPpm`,